Add files via upload
In these scripts, the wrist pronation/supination angle is analyzed to determine when the wrist is rotating while screwing in a screw, and when the wrist is not rotating. These loops scan the angle of rotation to determine if the wrist rotation angle reaches a certain threshold, then stores the angle data in a multi column matrix. The data is then scanned again in the second loop to determine if the wrist reached a higher threshold, determining if the data in each column was truly a wrist rotation or just a secondary motion of the wrist. The columns that do not reach this higher threshold are then removed from the matrix. The wrist rotation angles are then placed in a chronological, single column vector.